Samler basis funksjonalitet i en felles base for mottatt og send. Setter begge til aktive (disablet før)
